Add Credentials trait
This is to ensure that once a credential is rotated, accessing its
components such as access key and secret key are consistent
