+++ native/src/main/scala/sqs4s/api/lo/ReceiveMessage.scala
@@ -29,7 +29,7 @@ case class ReceiveMessage[F[_]: Sync: Clock: Timer, T](
     SignedRequest.post[F](
       params,
       config.queue,
-      config.credential,
+      config.credentials,
       config.region
     ).flatMap(_.render)
   }
