+++ native/src/main/scala/sqs4s/api/lo/SendMessage.scala
@@ -44,7 +44,7 @@ case class SendMessage[F[_]: Sync: Clock: Timer, T](
     SignedRequest.post[F](
       params,
       config.queue,
-      config.credential,
+      config.credentials,
       config.region
     ).flatMap(_.render)
   }
