+++ native/src/main/scala/sqs4s/api/lo/SendMessageBatch.scala
@@ -86,7 +86,7 @@ case class SendMessageBatch[F[_]: Sync: Clock: Timer, T](
     SignedRequest.post[F](
       params,
       config.queue,
-      config.credential,
+      config.credentials,
       config.region
     ).flatMap(_.render)
   }
